# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export type CustomerStatus = typeof CustomerStatus[keyof typeof CustomerStatus];


export const CustomerStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export interface Customer {
  id: string;
  name: string;
  address: string;
  phone: string;
  planName: string;
  monthlyRate: number;
  status: CustomerStatus;
  /** @nullable */
  notes?: string | null;
  balance: number;
  createdAt: string;
}

export type CustomerDetailStatus = typeof CustomerDetailStatus[keyof typeof CustomerDetailStatus];


export const CustomerDetailStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export type TransactionType = typeof TransactionType[keyof typeof TransactionType];


export const TransactionType = {
  service: 'service',
  equipment: 'equipment',
  one_time: 'one_time',
  late_fee: 'late_fee',
  manual_late_fee: 'manual_late_fee',
  payment: 'payment',
} as const;

export interface Transaction {
  id: string;
  customerId: string;
  date: string;
  type: TransactionType;
  description: string;
  amount: number;
  createdAt: string;
}

export interface CustomerDetail {
  id: string;
  name: string;
  address: string;
  phone: string;
  planName: string;
  monthlyRate: number;
  status: CustomerDetailStatus;
  /** @nullable */
  notes?: string | null;
  balance: number;
  createdAt: string;
  transactions: Transaction[];
}

export type CustomerInputStatus = typeof CustomerInputStatus[keyof typeof CustomerInputStatus];


export const CustomerInputStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export interface CustomerInput {
  /** @minLength 1 */
  name: string;
  address: string;
  phone: string;
  planName: string;
  /** @minimum 0 */
  monthlyRate: number;
  status?: CustomerInputStatus;
  notes?: string;
}

export type CustomerUpdateStatus = typeof CustomerUpdateStatus[keyof typeof CustomerUpdateStatus];


export const CustomerUpdateStatus = {
  active: 'active',
  inactive: 'inactive',
  suspended: 'suspended',
} as const;

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  phone?: string;
  planName?: string;
  /** @minimum 0 */
  monthlyRate?: number;
  status?: CustomerUpdateStatus;
  notes?: string;
}

export type TransactionInputType = typeof TransactionInputType[keyof typeof TransactionInputType];


export const TransactionInputType = {
  service: 'service',
  equipment: 'equipment',
  one_time: 'one_time',
  late_fee: 'late_fee',
  manual_late_fee: 'manual_late_fee',
  payment: 'payment',
} as const;

export interface TransactionInput {
  type: TransactionInputType;
  /** @minLength 1 */
  description: string;
  amount: number;
  date: string;
}

export type LateFeeResultDetailsItem = {
  customerId: string;
  customerName: string;
  balance: number;
  feeAmount: number;
};

export interface LateFeeResult {
  applied: number;
  skipped: number;
  totalFeesAdded: number;
  details?: LateFeeResultDetailsItem[];
}

export interface SpreadsheetInfo {
  spreadsheetId: string;
  url: string;
  title: string;
}

export interface RecentTransaction {
  id: string;
  customerId: string;
  customerName: string;
  date: string;
  type: string;
  description: string;
  amount: number;
}

export interface DashboardSummary {
  totalCustomers: number;
  activeCustomers: number;
  totalOutstanding: number;
  customersWithBalance: number;
  customersWithNegativeBalance: number;
  recentTransactions: RecentTransaction[];
}