# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerStatus } from './customerStatus';

export interface Customer {
  id: string;
  name: string;
  address: string;
  phone: string;
  planName: string;
  monthlyRate: number;
  status: CustomerStatus;
  /** @nullable */
  notes?: string | null;
  balance: number;
  createdAt: string;
}