# lib/api-zod/src/generated/types/customerUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerUpdateStatus } from './customerUpdateStatus';

export interface CustomerUpdate {
  /** @minLength 1 */
  name?: string;
  address?: string;
  phone?: string;
  planName?: string;
  /** @minimum 0 */
  monthlyRate?: number;
  status?: CustomerUpdateStatus;
  notes?: string;
}