# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { RecentTransaction } from './recentTransaction';

export interface DashboardSummary {
  totalCustomers: number;
  activeCustomers: number;
  totalOutstanding: number;
  customersWithBalance: number;
  customersWithNegativeBalance: number;
  recentTransactions: RecentTransaction[];
}