# lib/api-zod/src/generated/types/recentTransaction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface RecentTransaction {
  id: string;
  customerId: string;
  customerName: string;
  date: string;
  type: string;
  description: string;
  amount: number;
}