# lib/api-zod/src/generated/types/spreadsheetInfo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Internet Billing Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SpreadsheetInfo {
  spreadsheetId: string;
  url: string;
  title: string;
}